pull all travel specific style providers and themes to a separate file
